requirements.txt
# Core
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.3.2

# AI/ML
autogluon.tabular==1.5.0
catboost==1.2.8
xgboost==3.1.3
lightgbm==4.6.0
torch==2.9.1

# UI & Data
streamlit==1.30.0
tushare==1.4.24
altair==5.0.0
tqdm==4.66.0
